backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
google-genai>=1.5.0
together>=1.4.0
rembg>=2.0.68
Pillow>=11.1.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
